+++ src/main/java/mekanism/client/sound/SoundHandler.java
@@ -110,7 +110,7 @@ public class SoundHandler {
     public static void playSound(SoundEvent sound) {
         //TODO: this previously was called getRecord, but it seems to match up closer param wise to master than record
         // this should be double checked this works properly
-        playSound(SimpleSound.master(sound, 1.0F, (float) MekanismConfig.client.baseSoundVolume.get()));
+        playSound(SimpleSound.master(sound, 1.0F, MekanismConfig.client.baseSoundVolume.get()));
     }
 
     public static void playSound(ISound sound) {
